Toast message and fixed language changed
added toast message after user manage the modules

+++ php/blantern.php
@@ -15,11 +15,11 @@ if($method === "add")
         $receiptdate = date("Y-m-d", strtotime($receiptdate));
         $sql = "INSERT INTO BLantern(BLantern_id, member_eng_name, member_chi_name, contact_num, blessing_price, votive_price, breceipt_num, vreceipt_num,     receipt_date, price, remarks, recordedBy , recordedOn) VALUES ('".$_POST['id']."','".$_POST['english']."','".$_POST['chinese']."','".$_POST['contact']."','".$_POST['blessing_price']."','".$_POST['votive_price']."','".$_POST['blessing_receipt']."','".$_POST['votive_receipt']."','".$receiptdate."','".$_POST['amount']."','".$_POST['remarks']."', '".$_SESSION['name']."', '".date("Y-m-d H:i:s")."')";  
         if (mysqli_query($conn,$sql)) {
-            header("Location: ../transactions/create-blantern.php?name=transaction&aside=create-blantern&success=success");
+            header("Location: ../transactions/create-blantern.php?name=transaction&aside=create-blantern&success=success&lang=".$_SESSION['lang']."");
         } 
     }
     else{
-        header("Location: ../transactions/create-blantern.php?name=transaction&aside=create-blantern&success=fail");
+        header("Location: ../transactions/create-blantern.php?name=transaction&aside=create-blantern&success=fail&lang=".$_SESSION['lang']."");
     }
     
 
@@ -31,12 +31,20 @@ if($method === "delete")
 
     $id = $_GET["Id"];
    
-    $sql = "DELETE FROM BLantern WHERE BLantern_id ='".$_GET['Id']."'"; 
+    $mid_sql = "SELECT EXISTS (SELECT * FROM BLantern WHERE BLantern_id = '$id') AS row";
+    $result = $conn->query($mid_sql);  
+    $row = mysqli_fetch_array($result); 
+    if($row['row'] == 0){ 
+        header("Location: ../transactions/search-blantern.php?name=transaction&aside=search-blantern&status=fail&lang=".$_SESSION['lang']."");
+    }
+    else 
+    {
 
-    $result = $conn->query($sql);
-    header("Location: ../transactions/search-blantern.php?name=transaction&aside=search-blantern");
-    
-   
+        $sql = "DELETE FROM BLantern WHERE BLantern_id ='".$_GET['Id']."'"; 
+
+        $result = $conn->query($sql);
+        header("Location: ../transactions/search-blantern.php?name=transaction&aside=search-blantern&status=success&lang=".$_SESSION['lang']."");
+    }
 }
 
 
@@ -51,7 +59,7 @@ if($method === "update")
 
     $sql = "UPDATE BLantern SET  member_eng_name = '".$_POST['english']."', member_chi_name = '".$_POST['chinese']."', contact_num = '".$_POST['contact']."', blessing_price= '".$_POST['blessing_price']."', votive_price= '".$_POST['votive_price']."', breceipt_num = '".$_POST['blessing_receipt']."', vreceipt_num = '".$_POST['votive_receipt']."', receipt_date = '".$receiptdate."', price = '".$_POST['amount']."', remarks = '".$_POST['remarks']."' WHERE BLantern_id = '".$id."'";   
     $result = $conn->query($sql);
-    header("Location: ../transactions/edit-blantern.php?name=transaction&Id=".$id."&method=update");
+    header("Location: ../transactions/edit-blantern.php?name=transaction&Id=".$id."&method=update&status=success&lang=".$_SESSION['lang']."");
 
 }
